# Optimize `allocator` by removing `libc`.