By not using `libc::backtrace(` anymore, and replacing it with `backtrace::trace` we can reduce the overhead.

Before changes
```
alloc_32                time:   [30.665 ns 30.675 ns 30.684 ns]                                                                                                                                                    
alloc_32                time:   [31.007 ns 31.013 ns 31.020 ns]      
```

After doing the optimization

```
alloc_32                time:   [24.328 ns 24.342 ns 24.355 ns]                      
alloc_32                time:   [24.455 ns 24.466 ns 24.478 ns]     
```